Create a directory for subcommands
