+++ src/cli/cmd/mod.rs
@@ -1,0 +1,6 @@
+// --------
+// COMMANDS
+// --------
+
+mod view;
+pub use view::*;
